requirements_py2.txt
numpy==1.12.1
pytz==2015.4
more-itertools==5.0.0
zipp==0.6.0
cycler==0.10.0
matplotlib>=1.5.1,<3
pyparsing>=1.5.7,<3
backports.functools_lru_cache==1.6.4
future==0.18.2
tblib>=1.7.0
pg8000<1.13
websocket-client<1
user_agent<0.1.10
selectors==0.0.14
docker<5
pyyaml<6
ruamel.yaml<0.15.77
ruamel.yaml.clib<0.2.3
networkx==2.2
decorator<5
typing-extensions<4
pathlib2==2.3.2
kiwisolver<1.2
pyparsing<3
cwltool==1.0.20190618201008
contextlib2<1
importlib_metadata==1.7.0
packaging<21
pytest>=4.6.4,<4.7
configparser<5
pandas>=0.17.1,<0.23
futures>=3.2.0
